JobStatus/requirements.txt
requests
python-openstackclient
pyjwt
openstack-heat
